Prettify .json files with Jekyll-Tidy-JSON plugin
Contrary to json_pp, Jekyll-Tidy-JSON preserves order of entries
in JSONs it processes, which is good for diffing.  And since these JSONs
are no longer built with Liquid templates, we don't have to worry about
syntax errors which need to be fixed with json_pp.

JSON-LDs are still cleaned-up with json_pp.